+++ pignacio_scripts/nagios/__init__.py
@@ -17,5 +17,4 @@
 
 from .logger import NagiosLogger
 
-
 __all__ = ['NagiosLogger']